requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
beautifulsoup4=4.10.0=pyh06a4308_0
blas=1.0=mkl
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h9ed2024_1003
bs4=4.10.0=hd3eb1b0_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.10.8=py39hecd8cb5_0
cffi=1.14.6=py39h2125817_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=35.0.0=py39h2fd3fbb_0
cycler=0.11.0=pyhd3eb1b0_0
ffmpeg=4.3=h0a44026_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=3.2=pyhd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
jpeg=9d=h9ed2024_0
kiwisolver=1.3.1=py39h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.0=hacca55c_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
matplotlib=3.5.0=py39hecd8cb5_0
matplotlib-base=3.5.0=py39h4f681db_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_2
nettle=3.7.3=h230ac6f_1
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1l=h9ed2024_0
packaging=21.3=pyhd3eb1b0_0
pillow=8.4.0=py39h98e4679_0
pip=21.2.4=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.7=h88f2d9e_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.0=py3.9_0
readline=8.1=h9ed2024_0
requests=2.26.0=pyhd3eb1b0_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
tk=8.6.11=h7bc2e8c_0
torchvision=0.11.1=py39_cpu
tornado=6.1=py39h9ed2024_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
